checking token and Admin role

+++ src/api/resolvers/barrelResolver.ts
@@ -1,6 +1,9 @@
 import barrelModel from "../model/barrelModel";
 import {Loans} from "../../interfaces/Loans";
 import {Barrels} from "../../interfaces/Barrels";
+import {UserIdWithToken} from "../../interfaces/User";
+import {GraphQLError} from "graphql/index";
+import loanModel from "../model/loanModel";
 
 export default {
   Loans: {
@@ -17,18 +20,44 @@ export default {
     },
   },
   Mutation: {
-    createBarrel: async (_parent: unknown, args: {id: number}) => {
+    createBarrel: async (_parent: unknown, args: {id: number}, user: UserIdWithToken) => {
+      if(!user.token) {
+        throw new GraphQLError('User not logged in', {
+          extensions: {code: 'UNAUTHORIZED'},
+        });
+      }
+      if(user.role !== 'admin') {
+        throw new GraphQLError('Not ADMIN', {
+          extensions: {code: 'UNAUTHORIZED'},
+        });
+      }
       const barrel = new barrelModel(args);
       return barrel.save();
     },
-    updateBarrel: async (_parent: unknown, args: {id: number}) => {
-      //TODO: check token and if user can update barrel
-
+    updateBarrel: async (_parent: unknown, args: {id: number}, user: UserIdWithToken) => {
+      if(!user.token) {
+        throw new GraphQLError('User not logged in', {
+          extensions: {code: 'UNAUTHORIZED'},
+        });
+      }
+      if(user.role !== 'admin') {
+        throw new GraphQLError('Not ADMIN', {
+          extensions: {code: 'UNAUTHORIZED'},
+        });
+      }
       return barrelModel.findByIdAndUpdate(args.id, args, {new: true});
     },
-    deleteBarrel: async (_parent: unknown, args: {id: number}) => {
-      //TODO: check token and if user can delete barrel
-
+    deleteBarrel: async (_parent: unknown, args: {id: number}, user: UserIdWithToken) => {
+      if(!user.token) {
+        throw new GraphQLError('User not logged in', {
+          extensions: {code: 'UNAUTHORIZED'},
+        });
+      }
+      if(user.role !== 'admin') {
+        throw new GraphQLError('Not ADMIN', {
+          extensions: {code: 'UNAUTHORIZED'},
+        });
+      }
       return barrelModel.findByIdAndDelete(args.id);
     }
   }
